+++ packages/ui/src/2_molecules/VerticalTabsMobile/VerticalTabsMobile.types.ts
@@ -1,0 +1,23 @@
+import { ReactNode } from 'react';
+
+export type VerticalTabsMobileProps = {
+  items: VerticalTabsMobileItem[];
+  selectedIndex?: number | null;
+  className?: string;
+  tabsClassName?: string;
+  contentClassName?: string;
+  onChange?: (index: number | null) => void;
+  header?: (props: VerticalTabsMobileProps) => ReactNode;
+};
+
+export type VerticalTabsMobileItem = {
+  label: ReactNode;
+  content: ReactNode;
+  infoText?: ReactNode;
+  disabled?: boolean;
+  dataLayoutId?: string;
+};
+
+export type VerticalTabMobileItemButtonProps = VerticalTabsMobileItem & {
+  onClick: () => void;
+};